Refactor package layout under retirement_planner
